pyproject.toml
[tool.poetry]
name = "north-pole-buddy"
version = "0.1.0"
description = ""
authors = ["Adam Guillot <adam.guillot@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
mangum = "^0.17.0"
python-telegram-bot = "^20.7"
loguru = "^0.7.2"
openai = "^1.3.7"
boto3 = "^1.33.10"


[tool.poetry.group.nolambda.dependencies]
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
